StackOverflowClone/requirements.txt
asgiref==3.6.0
arrow==1.2.3
black==23.3.0
boto3==1.26.123
botocore==1.29.123
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap4==2022.1
cryptography==40.0.2
channels==3.0.4
channels-redis==4.1.0
defusedxml==0.7.1
django-resized==1.0.2
Django==3.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-dotenv==1.4.2
django-elasticsearch-dsl==7.3
django-environ==0.10.0
django-js-asset==2.0.0
django-storages==1.13.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
Markdown==3.4.3
oauthlib==3.2.2
openai==0.27.7
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
uvicorn===0.22.0